fake_currency-main (workinng webapp)/requirements.txt
keras==2.7.0
numpy==1.20.1
Pillow==9.4.0
streamlit==1.5.1
tensorflow
